Fix SQL injection in WP_Query 'm' date parameter
Add (int) cast to all substr() calls extracting date parts from the
'm' query parameter in WP_Query::get_posts(). Without the cast,
non-numeric characters in the parameter value could bypass quoting and
inject arbitrary SQL.

Includes 5 unit tests verifying:
- SQL keywords in 'm' are neutralized (UNION SELECT, OR 1=1)
- Non-numeric month fragment safely casts to 0
- Non-numeric prefix + SQL keywords produce empty results
- Normal datetime strings still work (regression check)

Props: @opencode
Fixes: #62161

+++ src/wp-includes/class-wp-query.php
@@ -2081,21 +2081,21 @@ class WP_Query {
 		}
 		// The "m" parameter is meant for months but accepts datetimes of varying specificity.
 		if ( $query_vars['m'] ) {
-			$where .= " AND YEAR({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 0, 4 );
+			$where .= " AND YEAR({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 0, 4 );
 			if ( strlen( $query_vars['m'] ) > 5 ) {
-				$where .= " AND MONTH({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 4, 2 );
+				$where .= " AND MONTH({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 4, 2 );
 			}
 			if ( strlen( $query_vars['m'] ) > 7 ) {
-				$where .= " AND DAYOFMONTH({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 6, 2 );
+				$where .= " AND DAYOFMONTH({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 6, 2 );
 			}
 			if ( strlen( $query_vars['m'] ) > 9 ) {
-				$where .= " AND HOUR({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 8, 2 );
+				$where .= " AND HOUR({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 8, 2 );
 			}
 			if ( strlen( $query_vars['m'] ) > 11 ) {
-				$where .= " AND MINUTE({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 10, 2 );
+				$where .= " AND MINUTE({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 10, 2 );
 			}
 			if ( strlen( $query_vars['m'] ) > 13 ) {
-				$where .= " AND SECOND({$wpdb->posts}.post_date)=" . substr( $query_vars['m'], 12, 2 );
+				$where .= " AND SECOND({$wpdb->posts}.post_date)=" . (int) substr( $query_vars['m'], 12, 2 );
 			}
 		}
